+++ src/org/sosy_lab/cpachecker/cfa/CfaMetadata.java
@@ -11,6 +11,7 @@ package org.sosy_lab.cpachecker.cfa;
 import static com.google.common.base.Preconditions.checkNotNull;
 
 import com.google.common.base.MoreObjects;
+import com.google.common.base.Preconditions;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableListMultimap;
 import java.nio.file.Path;
@@ -21,6 +22,7 @@ import org.checkerframework.checker.nullness.qual.Nullable;
 import org.sosy_lab.cpachecker.cfa.ast.acsl.ACSLAnnotation;
 import org.sosy_lab.cpachecker.cfa.model.CFAEdge;
 import org.sosy_lab.cpachecker.cfa.model.FunctionEntryNode;
+import org.sosy_lab.cpachecker.cfa.model.svlib.SvLibCfaMetadata;
 import org.sosy_lab.cpachecker.cfa.types.MachineModel;
 import org.sosy_lab.cpachecker.util.LiveVariables;
 import org.sosy_lab.cpachecker.util.LoopStructure;
@@ -45,6 +47,7 @@ public final class CfaMetadata {
   private final @Nullable VariableClassification variableClassification;
   private final @Nullable LiveVariables liveVariables;
   private final @Nullable ImmutableListMultimap<CFAEdge, ACSLAnnotation> edgesToAnnotations;
+  private final @Nullable SvLibCfaMetadata svLibCfaMetadata;
 
   private final @Nullable CfaTransformationMetadata transformationMetadata;
 
@@ -60,6 +63,7 @@ public final class CfaMetadata {
       @Nullable VariableClassification pVariableClassification,
       @Nullable LiveVariables pLiveVariables,
       @Nullable ImmutableListMultimap<CFAEdge, ACSLAnnotation> pEdgesToAnnotations,
+      @Nullable SvLibCfaMetadata pSvLibCfaMetadata,
       @Nullable CfaTransformationMetadata pCfaTransformationMetadata) {
     machineModel = checkNotNull(pMachineModel);
     cfaLanguage = checkNotNull(pCFALanguage);
@@ -73,7 +77,7 @@ public final class CfaMetadata {
     variableClassification = pVariableClassification;
     liveVariables = pLiveVariables;
     edgesToAnnotations = pEdgesToAnnotations;
-
+    svLibCfaMetadata = pSvLibCfaMetadata;
     transformationMetadata = pCfaTransformationMetadata;
   }
 
@@ -112,6 +116,7 @@ public final class CfaMetadata {
         null,
         null,
         null,
+        null,
         null);
   }
 
@@ -145,6 +150,7 @@ public final class CfaMetadata {
         variableClassification,
         liveVariables,
         edgesToAnnotations,
+        svLibCfaMetadata,
         transformationMetadata);
   }
 
@@ -162,6 +168,7 @@ public final class CfaMetadata {
             variableClassification,
             liveVariables,
             edgesToAnnotations,
+            svLibCfaMetadata,
             pTransformationMetadata);
     return newMetadata;
   }
@@ -228,6 +235,7 @@ public final class CfaMetadata {
         variableClassification,
         liveVariables,
         edgesToAnnotations,
+        svLibCfaMetadata,
         transformationMetadata);
   }
 
@@ -260,6 +268,7 @@ public final class CfaMetadata {
         variableClassification,
         liveVariables,
         edgesToAnnotations,
+        svLibCfaMetadata,
         transformationMetadata);
   }
 
@@ -305,6 +314,44 @@ public final class CfaMetadata {
         variableClassification,
         liveVariables,
         edgesToAnnotations,
+        svLibCfaMetadata,
+        transformationMetadata);
+  }
+
+  /**
+   * Returns the SV-LIB specific CFA metadata, if it's stored in this metadata instance.
+   *
+   * @return If this metadata instance contains the SV-LIB-specific CFA metadata, an optional
+   *     containing the SV-LIB-specific CFA metadata is returned. Otherwise, if this metadata
+   *     instance doesn't contain the SV-LIB-specific CFA metadata, an empty optional is returned.
+   */
+  public Optional<SvLibCfaMetadata> getSvLibCfaMetadata() {
+    return Optional.ofNullable(svLibCfaMetadata);
+  }
+
+  /**
+   * Returns a copy of this metadata instance, but with the specified SvLibCfaMetadata.
+   *
+   * @param pSvLibCfaMetadata the SvLibCfaMetadata to store in the returned metadata instance (use
+   *     {@code null} to create an instance without SvLibCfaMetadata)
+   * @return a copy of this metadata instance, but with the specified AST structure
+   */
+  public CfaMetadata withSvLibCfaMetadata(@Nullable SvLibCfaMetadata pSvLibCfaMetadata) {
+    Preconditions.checkArgument(
+        inputLanguage == Language.SVLIB ? pSvLibCfaMetadata != null : pSvLibCfaMetadata == null);
+    return new CfaMetadata(
+        machineModel,
+        cfaLanguage,
+        inputLanguage,
+        fileNames,
+        mainFunctionEntry,
+        connectedness,
+        astCFARelation,
+        loopStructure,
+        variableClassification,
+        liveVariables,
+        edgesToAnnotations,
+        pSvLibCfaMetadata,
         transformationMetadata);
   }
 
@@ -328,6 +375,7 @@ public final class CfaMetadata {
         variableClassification,
         liveVariables,
         edgesToAnnotations,
+        svLibCfaMetadata,
         transformationMetadata);
   }
 
@@ -363,6 +411,7 @@ public final class CfaMetadata {
         pVariableClassification,
         liveVariables,
         edgesToAnnotations,
+        svLibCfaMetadata,
         transformationMetadata);
   }
 
@@ -397,6 +446,7 @@ public final class CfaMetadata {
         variableClassification,
         pLiveVariables,
         edgesToAnnotations,
+        svLibCfaMetadata,
         transformationMetadata);
   }
 
@@ -435,6 +485,7 @@ public final class CfaMetadata {
         variableClassification,
         liveVariables,
         pedgesToAnnotations,
+        svLibCfaMetadata,
         transformationMetadata);
   }
 
@@ -451,6 +502,7 @@ public final class CfaMetadata {
         variableClassification,
         liveVariables,
         edgesToAnnotations,
+        svLibCfaMetadata,
         transformationMetadata);
   }
 
@@ -470,6 +522,8 @@ public final class CfaMetadata {
         && Objects.equals(variableClassification, other.variableClassification)
         && Objects.equals(liveVariables, other.liveVariables)
         && Objects.equals(edgesToAnnotations, other.edgesToAnnotations)
+        && Objects.equals(astCFARelation, other.astCFARelation)
+        && Objects.equals(svLibCfaMetadata, other.svLibCfaMetadata)
         && Objects.equals(transformationMetadata, other.transformationMetadata);
   }