+++ src/org/sosy_lab/cpachecker/cfa/ast/svlib/specification/SvLibFinalTerm.java
@@ -1,0 +1,82 @@
+// This file is part of CPAchecker,
+// a tool for configurable software verification:
+// https://cpachecker.sosy-lab.org
+//
+// SPDX-FileCopyrightText: 2025 Dirk Beyer <https://www.sosy-lab.org>
+//
+// SPDX-License-Identifier: Apache-2.0
+
+package org.sosy_lab.cpachecker.cfa.ast.svlib.specification;
+
+import java.io.Serial;
+import org.sosy_lab.cpachecker.cfa.ast.FileLocation;
+import org.sosy_lab.cpachecker.cfa.ast.svlib.SvLibAstNodeVisitor;
+import org.sosy_lab.cpachecker.cfa.ast.svlib.SvLibExpressionVisitor;
+import org.sosy_lab.cpachecker.cfa.ast.svlib.SvLibIdTerm;
+import org.sosy_lab.cpachecker.cfa.ast.svlib.SvLibTermVisitor;
+import org.sosy_lab.cpachecker.cfa.types.svlib.SvLibType;
+
+public final class SvLibFinalTerm implements SvLibFinalRelationalTerm {
+
+  @Serial private static final long serialVersionUID = 5381549261475877405L;
+  private final FileLocation fileLocation;
+  private final SvLibIdTerm term;
+
+  public SvLibFinalTerm(FileLocation pFileLocation, SvLibIdTerm pTerm) {
+    fileLocation = pFileLocation;
+    term = pTerm;
+  }
+
+  @Override
+  public SvLibType getExpressionType() {
+    return term.getExpressionType();
+  }
+
+  @Override
+  public FileLocation getFileLocation() {
+    return fileLocation;
+  }
+
+  @Override
+  public String toASTString(AAstNodeRepresentation pAAstNodeRepresentation) {
+    return "(old " + term.toASTString(pAAstNodeRepresentation) + ")";
+  }
+
+  @Override
+  public String toParenthesizedASTString(AAstNodeRepresentation pAAstNodeRepresentation) {
+    return toASTString(pAAstNodeRepresentation);
+  }
+
+  @Override
+  public <R, X extends Exception> R accept(SvLibAstNodeVisitor<R, X> v) throws X {
+    return v.accept(this);
+  }
+
+  @Override
+  public <R, X extends Exception> R accept(SvLibTermVisitor<R, X> v) throws X {
+    return v.accept(this);
+  }
+
+  @Override
+  public <R, X extends Exception> R accept(SvLibExpressionVisitor<R, X> v) throws X {
+    return v.accept(this);
+  }
+
+  @Override
+  public int hashCode() {
+    return term.hashCode();
+  }
+
+  @Override
+  public boolean equals(Object pO) {
+    if (this == pO) {
+      return true;
+    }
+
+    return pO instanceof SvLibFinalTerm other && term.equals(other.term);
+  }
+
+  public SvLibIdTerm getTerm() {
+    return term;
+  }
+}